refactor structrure
