+++ zulassung2/src/main/java/mops/zulassung2/model/dataObjects/AccountCreator.java
@@ -1,4 +1,4 @@
-package mops.zulassung2.model;
+package mops.zulassung2.model.dataObjects;
 
 import org.keycloak.KeycloakPrincipal;
 import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
