+++ zulassung2/src/main/java/mops/zulassung2/model/fileparsing/CustomCSVLineParser.java
@@ -1,6 +1,6 @@
 package mops.zulassung2.model.fileparsing;
 
-import mops.zulassung2.model.Student;
+import mops.zulassung2.model.dataObjects.Student;
 import org.apache.commons.csv.CSVRecord;
 
 public class CustomCSVLineParser implements CSVLineParser {
