+++ zulassung2/src/main/java/mops/zulassung2/services/OrganisatorService.java
@@ -1,6 +1,6 @@
 package mops.zulassung2.services;
 
-import mops.zulassung2.model.Student;
+import mops.zulassung2.model.dataObjects.Student;
 import mops.zulassung2.model.fileparsing.CustomCSVLineParser;
 import mops.zulassung2.model.fileparsing.CustomValidator;
 import mops.zulassung2.model.fileparsing.FileParser;
