+++ zulassung2/src/main/java/mops/zulassung2/services/ReceiptData.java
@@ -1,6 +1,6 @@
 package mops.zulassung2.services;
 
-import mops.zulassung2.model.Student;
+import mops.zulassung2.model.dataObjects.Student;
 
 public interface ReceiptData {
   String create(Student student, String currentSubject);
