+++ zulassung2/src/main/resources/templates/error-403.html
@@ -24,7 +24,7 @@
 <body>
 <div class="text-center">
     <h1>Sie haben nicht die benötigten Rechte</h1>
-    <img th:src="@{/yellow-sad-smiley.png}" class="img-fluid">
+    <img th:src="@{/images/yellow-sad-smiley.png}" class="img-fluid">
 </div>
 <div class="text-center">
     <h2><a th:href="@{/zulassung2/}">Hier gehts zurück</a></h2>
