+++ zulassung2/src/main/resources/templates/error-500.html
@@ -24,7 +24,7 @@
 <body>
 <div class="text-center">
     <h1>Es tut uns Leid ein unerwarteter Fehler ist unterlaufen Seite konnte leider nicht gefunden werden</h1>
-    <img th:src="@{/yellow-sad-smiley.png}" class="img-fluid">
+    <img th:src="@{/images/yellow-sad-smiley.png}" class="img-fluid">
 </div>
 <div class="text-center">
     <h2><a th:href="@{/zulassung2/}">Hier gehts zurück</a></h2>
