+++ zulassung2/src/main/resources/templates/orga.html
@@ -12,7 +12,7 @@
     <img alt="picture" src="images/orga-picture.jpg">
 </div>
 <div class="jumbotron" id="content" th:fragment="mycontent">
-    <link href="../static/css/stylesheet.css" rel="stylesheet" th:href="@{/css/stylesheet.css}">
+    <link href="../css/stylesheet.css" rel="stylesheet" th:href="@{/css/stylesheet.css}">
     <h1>Organisatoren</h1>
     <div>Hier können Sie die entsprechenden Dateien hochladen!</div>
     <div class="form-group" id="mops-upload-csv">
